Another relocation... documentation updates
