+++ src/main/java/com/e_gineering/collectd/Attribute.java
@@ -1,4 +1,4 @@
-package com.e_gineering;
+package com.e_gineering.collectd;
 
 import org.collectd.api.DataSet;
